+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * feat(addons): add maintenance windows manipulation with the new `addon-config` command ([PR#955](https://github.com/Scalingo/cli/pull/955))
 * feat(install.sh): verify the archive checksum ([PR#988](https://github.com/Scalingo/cli/pull/988))
 * feat(region): more debug logs  ([PR#1007](https://github.com/Scalingo/cli/pull/1008))
+* fix(events): `link_scm` data types
 
 ### 1.29.1